Remove hardcoded filter parameters in collection function
The hardcoded parameters for the API used to fetch collections have been replaced with a more generic URL in the collection function of lib.rs. This improves flexibility as it allows any anime collection request to be made, not just a specific one, and it improves code cleanliness. With this change, a client can specify their own parameters.